model/pyb/requirements.txt
talib-binary
datetime
numpy
alpaca_trade_api
pandas
matplotlib
pybuilder